backendapp/requirements.txt
fastapi>=0.110.0
uvicorn[standard]
pydantic
python-dotenv
pytest
httpx
python-multipart>=0.0.9
pdf2image>=1.17.0
paddlepaddle>=3.2.0
paddleocr>=3.4.0
ndlocr-lite @ git+https://github.com/ndl-lab/ndlocr-lite.git
python-Levenshtein>=0.25.0
jiwer>=3.0.0
